refactor: option 패키지명 변경 -> variant (#15)
