+++ src/main/java/com/kt/domain/variant/VariantType.java
@@ -1,11 +1,11 @@
-package com.kt.domain.option;
+package com.kt.domain.variant;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
 @Getter
 @RequiredArgsConstructor
-public enum OptionType {
+public enum VariantType {
 	SIZE("사이즈"),
 	COLOR("색");
