+++ src/main/java/com/kt/repository/variant/VariantRepository.java
@@ -2,7 +2,7 @@ package com.kt.repository.variant;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.kt.domain.option.Variant;
+import com.kt.domain.variant.Variant;
 
 public interface VariantRepository extends JpaRepository<Variant, Long> {
 }
